pyproject.toml
[tool.poetry]
name = "hw-asyncio"
version = "0.1.0"
description = ""
authors = ["Maksim Molianov <mm_89615@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
requests = "^2.32.3"
aiohttp = "^3.11.8"
more-itertools = "^10.5.0"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
